+++ candid/src/bindings/mod.rs
@@ -1,1 +1,2 @@
+pub mod analysis;
 pub mod javascript;
