+++ candid/tests/assets/ok/escape.js
@@ -1,0 +1,9 @@
+({ IDL }) => {
+  const t = IDL.Record({
+    '\"' : IDL.Nat,
+    '\'' : IDL.Nat,
+    '\"\'' : IDL.Nat,
+    '\\\n\'\"' : IDL.Nat
+  });
+  return IDL.Service({'\n\'\"\'\'\"\"\r\t' : IDL.Func([t], [], [])});
+}
