+++ candid/tests/assets/ok/surrogate.fail
@@ -1,0 +1,1 @@
+Candid parser error: Unicode escape out of range d800
