+++ candid/tests/assets/surrogate.did
@@ -1,0 +1,3 @@
+type A = record {
+ "\u{D800}":nat;
+};
