+++ candid/tests/assets/unicode.did
@@ -1,0 +1,16 @@
+type A = record {
+ "字段名":nat;
+ "字 段 名2":nat;
+ "📦🍦":nat;
+ "\u{E000}":nat;
+};
+type B = variant { ""; "空的"; "  空的  "; "1⃣️2⃣️3⃣️"; };
+
+service C:
+{
+  "函数名": (A) -> (B);
+  "": (nat) -> (nat);
+  "👀": (nat) -> (nat) query;
+  "✈️  🚗 ⛱️ ": () -> () oneway;
+}
+
